+++ example/lib/main.dart
@@ -5,6 +5,7 @@ import 'network_image_demo.dart';
 import 'sliver_demo.dart';
 import 'sliver_easy_refresh_demo.dart';
 import 'tab_bar_demo.dart';
+import 'tags_demo.dart';
 import 'text_button_demo.dart';
 import 'text_field_demo.dart';
 
@@ -36,6 +37,7 @@ Map<String, Widget> routes = {
   "TabBar": const TabBarDemo(),
   "Sliver": const SliverDemo(),
   "Sliver EasyRefresh": const SliverEasyRefreshDemo(),
+  "Tags": const TagsDemo(),
 };
 
 class DemoPage extends StatelessWidget {