+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.1
 
 * fixed: 修复WrapperEasyRefresh的header和footer属性不生效问题
+* feat: WrapperTextButton增加属性：overflow, maxLines
 
 ## 1.1.0