+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.0
+
+* feat: 新增WrapperTags组件
+
 ## 1.1.1
 
 * fixed: 修复WrapperEasyRefresh的header和footer属性不生效问题